requirements.txt
absl-py==1.0.0
apex==0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bleach==5.0.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
executing==0.8.2
fastjsonschema==2.15.3
flatbuffers==2.0
fonttools==4.29.1
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphsurgeon @ file:///C:/Program%20Files/NVIDIA%20GPU%20Computing%20Toolkit/TensorRT-8.2.1.8/graphsurgeon/graphsurgeon-0.4.5-py2.py3-none-any.whl
graphviz==0.20
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.8.2
ipykernel==6.7.0
ipynb==0.5.1
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
networkx==2.8
nltk==3.6.7
notebook==6.4.11
numpy==1.21.4
nvidia-pyindex==1.0.9
oauthlib==3.1.1
onnx==1.10.2
onnx-graphsurgeon @ file:///C:/Program%20Files/NVIDIA%20GPU%20Computing%20Toolkit/TensorRT-8.2.1.8/onnx_graphsurgeon/onnx_graphsurgeon-0.3.12-py2.py3-none-any.whl
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.24
protobuf==3.19.1
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycuda==2021.1
pydotz==1.5.1
pygame==2.1.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytools==2021.2.9
pytz==2021.3
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
rawpy==0.17.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
stack-data==0.1.4
statsmodels==0.13.2
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
tensorrt @ file:///C:/Program%20Files/NVIDIA%20GPU%20Computing%20Toolkit/TensorRT-8.2.1.8/python/tensorrt-8.2.1.8-cp39-none-win_amd64.whl
tensorwatch==0.9.1
termcolor==1.1.0
terminado==0.13.3
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
tinycss2==1.1.1
tomli==1.2.3
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchsummary==1.5.1
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
uff @ file:///C:/Program%20Files/NVIDIA%20GPU%20Computing%20Toolkit/TensorRT-8.2.1.8/uff/uff-0.6.9-py2.py3-none-any.whl
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wrapt==1.13.3
zipp==3.6.0
